Restructure tree
This changeset allows for specific commit points of the esp-idf
framework to be referenced, so all builds are consistent. Initially we
are targetting the current master (head-of-tree) of the esp-idf
framework. However, if needed, we can decide when to track the esp-idf
tree (i.e. a specific "stable" release point, or the latest to get
important bug-fixes or feature changes. If needed we can also create a
local fork of the esp-idf, but that should hopefully be unlikely.